Add new unit spec covering examples from RFC 9421 document
Its objective is to ensure thorough compliance of this
library with what is specified in RFC 9421.

This is a work in progress and for now only covers examples
from section 2 but the idea is to cover the whole RFC,
if possible.

It already helped to discover and fix a few bugs.